Moved KnoedlerFilePipeline into pipeline.projects.knoedler.
